+++ .github/jobs/test-used-linker.sh
@@ -1,6 +1,7 @@
 #!/usr/bin/env bash
 set -xeuo pipefail
 
+cargo clean
 RUSTDOCFLAGS='--cfg linktime_used_linker' \
     RUSTFLAGS='-D warnings --cfg linktime_used_linker' \
     cargo test --target "$TARGET"