+++ link-section/src/lib.rs
@@ -3,9 +3,13 @@
 #![doc = include_str!("../docs/PREAMBLE.md")]
 #![allow(unsafe_code)]
 #![cfg_attr(linktime_used_linker, doc(test(attr(feature(used_with_arg)))))]
+#![no_std]
 
 #[doc = include_str!("../docs/LIFE_BEFORE_MAIN.md")]
 pub mod life_before_main {}
+
+#[cfg(target_family = "wasm")]
+mod wasm;
 
 #[cfg(target_family = "wasm")]
 extern crate alloc;
@@ -46,6 +50,12 @@ pub mod __support {
         ) -> usize;
     }
 
+    #[cfg(target_family = "wasm")]
+    pub use crate::wasm::{register_wasm_link_section_item, LinkSectionRawInfo};
+
+    #[cfg(all(not(miri), target_vendor = "pc"))]
+    pub use section::Alignment;
+
     /// Declares the section_name macro.
     #[macro_export]
     #[doc(hidden)]
@@ -79,7 +89,7 @@ pub mod __support {
                 )*
                 ($pattern:tt $unknown_section:ident $unknown_type:ident $name:ident) => {
                     const _: () = {
-                        compile_error!("Unknown section type: `{}`/`{}`", stringify!($unknown_section), stringify!($unknown_type));
+                        compile_error!(concat!("Unknown section type: `", stringify!($unknown_section), "/", stringify!($unknown_type), "`"));
                     };
                 };
             }
@@ -151,10 +161,9 @@ pub mod __support {
             }
         };
         ($section:ident $type:ident $name:ident $($aux:ident)? #[$attr:ident = __]
-            $(#[$meta:meta])* $($item:tt)*) => {
+            $($item:tt)*) => {
             $crate::__add_section_link_attribute_impl!(
                 $section $type $name $($aux)? #[$attr = __]
-                $(#[$meta])*
                 #[allow(unsafe_code)]
                 $($item)*
             );
@@ -208,6 +217,7 @@ pub mod __support {
             data section => (".data", ".link_section.") __ ();
             code bare =>    (".text", ".link_section.") __ ();
             code section => (".text", ".link_section.") __ ();
+            data bounds =>   (".data", ".link_section.") __ (".bounds");
         }
         AUXILIARY = ".";
         MAX_LENGTH = 16;
@@ -356,7 +366,7 @@ pub mod __support {
                         let name = $crate::__section_name!(
                             raw data bare $ident
                         );
-                        unsafe { <$ty>::new(name, section.0, section.1) }
+                        unsafe { <$ty>::new(name, section) }
                     };
                     &SECTION
                 }
@@ -372,7 +382,7 @@ pub mod __support {
                         let name = $crate::__section_name!(
                             raw data bare $aux $ident // swap
                         );
-                        unsafe { <$ty>::new(name, section.0, section.1) }
+                        unsafe { <$ty>::new(name, section) }
                     };
                     &SECTION
                 }
@@ -386,13 +396,11 @@ pub mod __support {
         #[macro_export]
         macro_rules! __get_section {
             (name=$ident:ident, type=$generic_ty:ty, aux=$($aux:ident)?) => {{
-                (std::ptr::null_mut(), std::ptr::null_mut())
+                $crate::__support::PtrBounds::new(core::ptr::null_mut(), core::ptr::null_mut())
             }};
         }
 
-        /// On Apple platforms, the linker provides a pointer to the start and end
-        /// of the section regardless of the section's name.
-        pub type SectionPtr<T> = *const ::core::marker::PhantomData<T>;
+        pub type Bounds = crate::__support::PtrBounds;
     }
 
     #[cfg(all(not(miri), target_family = "wasm"))]
@@ -402,25 +410,27 @@ pub mod __support {
         macro_rules! __get_section {
             (name=$ident:ident, type=$generic_ty:ty, aux=$($aux:ident)?) => {
                 {
-                    static __START: ::core::sync::atomic::AtomicPtr::<::core::marker::PhantomData<$generic_ty>> = unsafe {
-                        ::core::sync::atomic::AtomicPtr::<::core::marker::PhantomData<$generic_ty>>::new(::core::ptr::null_mut())
-                    };
-                    static __END: ::core::sync::atomic::AtomicPtr::<::core::marker::PhantomData<$generic_ty>> = unsafe {
-                        ::core::sync::atomic::AtomicPtr::<::core::marker::PhantomData<$generic_ty>>::new(::core::ptr::null_mut())
-                    };
+                    static __LINK_SECTION_NAME: &'static str = $crate::__section_name!(
+                        raw data bare $ident $($aux)?
+                    );
+                    $crate::__support::add_section_link_attribute!(
+                        data bounds $ident $($aux)?
+                        #[export_name = __]
+                        #[used]
+                        static mut __LINK_SECTION_INFO: $crate::__support::LinkSectionRawInfo = $crate::__support::LinkSectionRawInfo::new::<$generic_ty>(__LINK_SECTION_NAME);
+                    );
 
-                    (&__START, &__END)
+                    unsafe { $crate::__support::Bounds::new(&raw mut __LINK_SECTION_INFO) }
                 }
             }
         }
 
-        /// On WASM, we use an atomic pointer to the start and end of the
-        /// section. The host environment is responsible for registering the
-        /// section with the runtime.
-        pub type SectionPtr<T> =
-            &'static ::core::sync::atomic::AtomicPtr<::core::marker::PhantomData<T>>;
+        pub use crate::wasm::Bounds;
     }
 
+    /// On Windows platforms we don't have start/end symbols, but we do have
+    /// section sorting so we drop a minimum-sized type with the same alignment
+    /// as T at the start and end of the section.
     #[cfg(all(not(miri), target_vendor = "pc"))]
     mod section {
         #[doc(hidden)]
@@ -428,31 +438,59 @@ pub mod __support {
         macro_rules! __get_section {
             (name=$ident:ident, type=$generic_ty:ty, aux=$($aux:ident)?) => {
                 {
-                    $crate::__support::add_section_link_attribute!(
+                    use $crate::__support::Alignment;
+                    use $crate::__support::PtrBounds;
+                    use $crate::__support::add_section_link_attribute;
+                    use core::mem;
+
+                    add_section_link_attribute!(
                         data start $ident $($aux)?
                         #[link_section = __]
-                        static __START: [$generic_ty; 0] = [];
+                        static __START: Alignment<$generic_ty> = Alignment::new();
                     );
-                    $crate::__support::add_section_link_attribute!(
+                    add_section_link_attribute!(
                         data end $ident $($aux)?
                         #[link_section = __]
-                        static __END: [$generic_ty; 0] = [];
+                        static __END: Alignment<$generic_ty> = Alignment::new();
                     );
 
-                    (
-                        unsafe { &raw const __START as $crate::__support::SectionPtr<$generic_ty> },
-                        unsafe { &raw const __END as $crate::__support::SectionPtr<$generic_ty> },
+                    PtrBounds::new(
+                        unsafe {
+                            let start = &raw const __START;
+                            start.cast::<u8>().add(mem::size_of::<Alignment<$generic_ty>>()) as *const()
+                        },
+                        unsafe { &raw const __END as *const () },
                     )
                 }
             }
         }
 
-        /// On Windows platforms we don't have start/end symbols, but we do have
-        /// section sorting so we drop a [T; 0] at the start and end of the
+        /// A non-zero-sized type that is used to align the start and end of the
         /// section.
-        pub type SectionPtr<T> = *const [T; 0];
+        #[repr(C)]
+        pub struct Alignment<T> {
+            _align: [T; 0],
+            _padding: u8,
+        }
+
+        #[allow(clippy::new_without_default)]
+        impl<T> Alignment<T> {
+            pub const fn new() -> Self {
+                Self {
+                    _align: [],
+                    _padding: 0,
+                }
+            }
+        }
+
+        pub type Bounds = crate::__support::PtrBounds;
     }
 
+    /// On LLVM/GCC platforms we can use orphan sections with _start and _end
+    /// symbols.
+    ///
+    /// On Apple platforms, the linker provides a pointer to the start and end
+    /// of the section regardless of the section's name.
     #[cfg(all(not(miri), not(target_family = "wasm"), not(target_vendor = "pc")))]
     mod section {
         #[doc(hidden)]
@@ -460,32 +498,33 @@ pub mod __support {
         macro_rules! __get_section {
             (name=$ident:ident, type=$generic_ty:ty, aux=$($aux:ident)?) => {
                 {
+                    // These are not valid items, but they are valid pointers.
+                    // We cannot safely use them - only take pointers to them.
                     $crate::__support::add_section_link_attribute!(
                         data start $ident $($aux)?
                         #[link_name = __]
                         extern "C" {
-                            static __START: $crate::__support::SectionPtr<$generic_ty>;
+                            static __START: u8;
                         }
                     );
                     $crate::__support::add_section_link_attribute!(
                         data end $ident $($aux)?
                         #[link_name = __]
                         extern "C" {
-                            static __END: $crate::__support::SectionPtr<$generic_ty>;
+                            static __END: u8;
                         }
                     );
 
-                    (
-                        unsafe { &raw const __START as $crate::__support::SectionPtr<$generic_ty> },
-                        unsafe { &raw const __END as $crate::__support::SectionPtr<$generic_ty> },
+                    $crate::__support::PtrBounds::new(
+                        // TODO: black_box when hint is stable
+                        unsafe { &raw const __START as *const () },
+                        unsafe { &raw const __END as *const () },
                     )
                 }
             }
         }
 
-        /// On LLVM/GCC platforms we can use orphan sections with _start and
-        /// _end symbols.
-        pub type SectionPtr<T> = *const ::core::marker::PhantomData<T>;
+        pub type Bounds = crate::__support::PtrBounds;
     }
 
     /// Export a symbol into a link section.
@@ -582,49 +621,80 @@ pub mod __support {
             $stored_ty
         };
         ($type_source:tt, $ident:ident, $($aux:ident)?, $path:path, ($(#[$meta:meta])* $vis:vis fn $ident_fn:ident($($args:tt)*) $(-> $ret:ty)? $body:block)) => {
-            const _: () = {
-                type __InSecStoredTy = $crate::__in_section_crate!(@type_select $type_source $path, fn($($args)*) $(-> $ret)?);
-                $crate::__add_section_link_attribute!(
-                    data section $ident $($aux)?
-                    #[link_section = __]
-                    $(#[$meta])*
-                    #[allow(non_upper_case_globals)]
-                    $vis static __LINK_SECTION_FN_ITEM: __InSecStoredTy = $ident_fn;
-                );
-            };
-
+            $crate::__in_section_crate!($type_source, $ident, $($aux)?, $path, (
+                const _: fn($($args)*) $(-> $ret)? = $ident_fn;
+            ));
             $crate::__add_section_link_attribute!(
                 code section $ident $($aux)?
                 #[link_section = __]
-                fn $ident_fn($($args)*) $(-> $ret)? $body
+                $(#[$meta])*
+                $vis fn $ident_fn($($args)*) $(-> $ret)? $body
             );
         };
         ($type_source:tt, $ident:ident, $($aux:ident)?, $path:path, ($(#[$meta:meta])* $vis:vis static _ : $ty:ty = $value:expr;)) => {
-            const _: () = {
-                type __InSecStoredTy = $crate::__in_section_crate!(@type_select $type_source $path, $ty);
-                $crate::__add_section_link_attribute!(
-                    data section $ident $($aux)?
-                    #[link_section = __]
-                    $(#[$meta])* $vis static ANONYMOUS: __InSecStoredTy = $value;
-                );
-            };
+            $crate::__in_section_crate!($type_source, $ident, $($aux)?, $path, (
+                $(#[$meta])*
+                const _: fn($($args)*) $(-> $ret)? = $ident_fn;
+            ));
         };
         ($type_source:tt, $ident:ident, $($aux:ident)?, $path:path, ($(#[$meta:meta])* $vis:vis static $ident_static:ident : $ty:ty = $value:expr;)) => {
+            #[cfg(target_family = "wasm")]
+            compile_error!("static items are not supported on WASM: use const items instead");
+
+            #[cfg(not(target_family = "wasm"))]
             $crate::__add_section_link_attribute!(
                 data section $ident $($aux)?
                 #[link_section = __]
-                $(#[$meta])* $vis static $ident_static: $crate::__in_section_crate!(@type_select $type_source $path, $ty) = $value;
+                $(#[$meta])*
+                $vis static $ident_static: $crate::__in_section_crate!(@type_select $type_source $path, $ty) = $value;
             );
         };
         ($type_source:tt, $ident:ident, $($aux:ident)?, $path:path, ($(#[$meta:meta])* $vis:vis const $name:ident: $ty:ty = $value:expr;)) => {
             $(#[$meta])* $vis const $name: $ty = {
                 type __InSecStoredTy = $crate::__in_section_crate!(@type_select $type_source $path, $ty);
                 const __LINK_SECTION_CONST_ITEM_VALUE: __InSecStoredTy = $value;
+
+                #[cfg(target_family = "wasm")]
+                {
+                    // Register a counting item
+                    $crate::__add_section_link_attribute!(
+                        data section $ident $($aux)?
+                        #[link_section = __]
+                        $vis static __LINK_SECTION_CONST_ITEM: u8 = 0;
+                    );
+
+                    $crate::__add_section_link_attribute!(
+                        data bounds $ident $($aux)?
+                        #[link_name = __]
+                        extern "C" {
+                            static mut __LINK_SECTION_INFO: $crate::__support::LinkSectionRawInfo;
+                        }
+                    );
+
+                    #[link_section = ".init_array.0"]
+                    static mut __LINK_SECTION_ITEM_FN_REF: extern "C" fn() = {
+                        extern "C" fn __LINK_SECTION_ITEM_FN() {
+                            static DISARMED: ::core::sync::atomic::AtomicBool = ::core::sync::atomic::AtomicBool::new(false);
+                            if DISARMED.swap(true, ::core::sync::atomic::Ordering::Relaxed) {
+                                return;
+                            }
+                            unsafe {
+                                let ptr = $crate::__support::register_wasm_link_section_item(&raw mut __LINK_SECTION_INFO);
+                                ::core::ptr::write(ptr as *mut __InSecStoredTy, __LINK_SECTION_CONST_ITEM_VALUE);
+                            }
+                        }
+                        __LINK_SECTION_ITEM_FN
+                    };
+                }
+
+                #[cfg(not(target_family = "wasm"))]
                 $crate::__add_section_link_attribute!(
                     data section $ident $($aux)?
                     #[link_section = __]
-                    $(#[$meta])* $vis static __LINK_SECTION_CONST_ITEM: __InSecStoredTy = __LINK_SECTION_CONST_ITEM_VALUE;
+                    $(#[$meta])*
+                    $vis static __LINK_SECTION_CONST_ITEM: __InSecStoredTy = __LINK_SECTION_CONST_ITEM_VALUE;
                 );
+
                 __LINK_SECTION_CONST_ITEM_VALUE
             };
         };
@@ -635,7 +705,8 @@ pub mod __support {
                 $crate::__add_section_link_attribute!(
                     data section $ident $($aux)?
                     #[link_section = __]
-                    $(#[$meta])* $vis static __LINK_SECTION_CONST_ITEM: __InSecStoredTy = $value;
+                    $(#[$meta])*
+                    $vis static __LINK_SECTION_CONST_ITEM: __InSecStoredTy = $value;
                 );
             };
         };
@@ -643,7 +714,8 @@ pub mod __support {
             $crate::__add_section_link_attribute!(
                 data section $ident $($aux)?
                 #[link_section = __]
-                $(#[$meta])* $item
+                $(#[$meta])*
+                $item
             );
         };
     }
@@ -656,7 +728,33 @@ pub mod __support {
         type Item = T;
     }
 
-    pub use section::SectionPtr;
+    pub struct PtrBounds {
+        pub start: *const (),
+        pub end: *const (),
+    }
+
+    impl PtrBounds {
+        pub const fn new(start: *const (), end: *const ()) -> Self {
+            Self { start, end }
+        }
+
+        #[inline(always)]
+        pub const fn start_ptr(&self) -> *const () {
+            self.start
+        }
+        #[inline(always)]
+        pub const fn end_ptr(&self) -> *const () {
+            self.end
+        }
+        #[inline(always)]
+        pub const fn byte_len(&self) -> usize {
+            // NOTE: MSRV for non-WASM targets doesn't allow byte_offset_from,
+            // so we manually implement it here.
+            unsafe { (self.end.cast::<u8>()).offset_from(self.start.cast::<u8>()) as usize }
+        }
+    }
+
+    pub use section::Bounds;
 }
 
 /// Define a link section.
@@ -702,99 +800,30 @@ pub use ::linktime_proc_macro::in_section;
 #[repr(C)]
 pub struct Section {
     name: &'static str,
-    start: __support::SectionPtr<()>,
-    end: __support::SectionPtr<()>,
+    bounds: __support::Bounds,
 }
 
 impl Section {
     #[doc(hidden)]
-    pub const unsafe fn new(
-        name: &'static str,
-        start: __support::SectionPtr<()>,
-        end: __support::SectionPtr<()>,
-    ) -> Self {
-        Self { name, start, end }
+    pub const unsafe fn new(name: &'static str, bounds: __support::Bounds) -> Self {
+        Self { name, bounds }
     }
 
     /// The byte length of the section.
+    #[inline]
     pub fn byte_len(&self) -> usize {
-        unsafe { (self.end_ptr() as *const u8).offset_from(self.start_ptr() as *const u8) as usize }
+        self.bounds.byte_len()
     }
-}
 
-#[cfg(target_family = "wasm")]
-impl Section {
     /// The start address of the section.
+    #[inline]
     pub fn start_ptr(&self) -> *const () {
-        let ptr = self.start.load(::core::sync::atomic::Ordering::Relaxed) as *const ();
-        if ptr.is_null() {
-            unsafe {
-                let required = crate::__support::read_custom_section(
-                    self.name.as_ptr(),
-                    self.name.len(),
-                    ::core::ptr::null_mut(),
-                    0,
-                );
-                let ptr = ::alloc::alloc::alloc(
-                    ::core::alloc::Layout::from_size_align(required, 1).unwrap(),
-                );
-                self.start
-                    .store(ptr as _, ::core::sync::atomic::Ordering::Relaxed);
-                self.end.store(
-                    ptr.add(required) as _,
-                    ::core::sync::atomic::Ordering::Relaxed,
-                );
-                crate::__support::read_custom_section(
-                    self.name.as_ptr(),
-                    self.name.len(),
-                    ptr as *mut u8,
-                    required,
-                );
-            }
-        }
-        ptr
+        self.bounds.start_ptr()
     }
     /// The end address of the section.
+    #[inline]
     pub fn end_ptr(&self) -> *const () {
-        let ptr = self.end.load(::core::sync::atomic::Ordering::Relaxed) as *const ();
-        if ptr.is_null() {
-            unsafe {
-                let required = crate::__support::read_custom_section(
-                    self.name.as_ptr(),
-                    self.name.len(),
-                    ::core::ptr::null_mut(),
-                    0,
-                );
-                let ptr = ::alloc::alloc::alloc(
-                    ::core::alloc::Layout::from_size_align(required, 1).unwrap(),
-                );
-                self.start
-                    .store(ptr as _, ::core::sync::atomic::Ordering::Relaxed);
-                self.end.store(
-                    ptr.add(required) as _,
-                    ::core::sync::atomic::Ordering::Relaxed,
-                );
-                crate::__support::read_custom_section(
-                    self.name.as_ptr(),
-                    self.name.len(),
-                    ptr as *mut u8,
-                    required,
-                );
-            }
-        }
-        ptr
-    }
-}
-
-#[cfg(not(target_family = "wasm"))]
-impl Section {
-    /// The start address of the section.
-    pub fn start_ptr(&self) -> *const () {
-        self.start as *const ()
-    }
-    /// The end address of the section.
-    pub fn end_ptr(&self) -> *const () {
-        self.end as *const ()
+        self.bounds.end_ptr()
     }
 }
 
@@ -817,110 +846,46 @@ unsafe impl Send for Section {}
 #[repr(C)]
 pub struct TypedSection<T: 'static> {
     name: &'static str,
-    start: __support::SectionPtr<T>,
-    end: __support::SectionPtr<T>,
+    bounds: __support::Bounds,
     _phantom: ::core::marker::PhantomData<T>,
 }
 
-#[cfg(target_family = "wasm")]
-impl<T: 'static> TypedSection<T> {
-    /// The start address of the section.
-    pub fn start_ptr(&self) -> *const T {
-        let ptr = self.start.load(::core::sync::atomic::Ordering::Relaxed) as *const T;
-        if ptr.is_null() {
-            unsafe {
-                let required = crate::__support::read_custom_section(
-                    self.name.as_ptr(),
-                    self.name.len(),
-                    ::core::ptr::null_mut(),
-                    0,
-                );
-                let ptr = ::alloc::alloc::alloc(
-                    ::core::alloc::Layout::from_size_align(required, ::core::mem::align_of::<T>())
-                        .unwrap(),
-                );
-                self.start
-                    .store(ptr as _, ::core::sync::atomic::Ordering::Relaxed);
-                self.end.store(
-                    ptr.add(required) as _,
-                    ::core::sync::atomic::Ordering::Relaxed,
-                );
-                crate::__support::read_custom_section(
-                    self.name.as_ptr(),
-                    self.name.len(),
-                    ptr as *mut u8,
-                    required,
-                );
-            }
-        }
-        ptr
-    }
-
-    /// The end address of the section.
-    pub fn end_ptr(&self) -> *const T {
-        let ptr = self.end.load(::core::sync::atomic::Ordering::Relaxed) as *const T;
-        if ptr.is_null() {
-            unsafe {
-                let required = crate::__support::read_custom_section(
-                    self.name.as_ptr(),
-                    self.name.len(),
-                    ::core::ptr::null_mut(),
-                    0,
-                );
-                let ptr = ::alloc::alloc::alloc(
-                    ::core::alloc::Layout::from_size_align(required, ::core::mem::align_of::<T>())
-                        .unwrap(),
-                );
-                self.start
-                    .store(ptr as _, ::core::sync::atomic::Ordering::Relaxed);
-                self.end.store(
-                    ptr.add(required) as _,
-                    ::core::sync::atomic::Ordering::Relaxed,
-                );
-                crate::__support::read_custom_section(
-                    self.name.as_ptr(),
-                    self.name.len(),
-                    ptr as *mut u8,
-                    required,
-                );
-            }
-        }
-        ptr
-    }
-}
-
-#[cfg(not(target_family = "wasm"))]
-impl<T: 'static> TypedSection<T> {
-    /// The start address of the section.
-    #[inline(always)]
-    pub fn start_ptr(&self) -> *const T {
-        self.start as *const T
-    }
-
-    /// The end address of the section.
-    #[inline(always)]
-    pub fn end_ptr(&self) -> *const T {
-        self.end as *const T
-    }
-}
-
-// Non-const, shared functions (or functions that don't depend on the pointers)
 impl<T: 'static> TypedSection<T> {
     #[doc(hidden)]
-    pub const unsafe fn new(
-        name: &'static str,
-        start: __support::SectionPtr<T>,
-        end: __support::SectionPtr<T>,
-    ) -> Self {
+    pub const unsafe fn new(name: &'static str, bounds: __support::Bounds) -> Self {
         Self {
             name,
-            start,
-            end,
+            bounds,
             _phantom: ::core::marker::PhantomData,
         }
     }
 
+    /// The start address of the section.
+    #[inline(always)]
+    pub fn start_ptr(&self) -> *const T {
+        self.bounds.start_ptr() as *const T
+    }
+
+    /// The end address of the section.
+    #[inline(always)]
+    pub fn end_ptr(&self) -> *const T {
+        self.bounds.end_ptr() as *const T
+    }
+
+    /// The start address of the section.
+    #[inline]
+    pub fn start_ptr_mut(&self) -> *mut T {
+        self.bounds.start_ptr() as *mut T
+    }
+
+    /// The start address of the section.
+    #[inline]
+    pub fn end_ptr_mut(&self) -> *mut T {
+        self.bounds.end_ptr() as *mut T
+    }
+
     /// The stride of the typed section.
+    #[inline(always)]
     pub const fn stride(&self) -> usize {
         assert!(
             ::core::mem::size_of::<T>() > 0
@@ -930,21 +895,25 @@ impl<T: 'static> TypedSection<T> {
     }
 
     /// The byte length of the section.
+    #[inline]
     pub fn byte_len(&self) -> usize {
-        unsafe { (self.end_ptr() as *const u8).offset_from(self.start_ptr() as *const u8) as usize }
+        self.bounds.byte_len()
     }
 
     /// The number of elements in the section.
+    #[inline]
     pub fn len(&self) -> usize {
         self.byte_len() / self.stride()
     }
 
     /// True if the section is empty.
+    #[inline]
     pub fn is_empty(&self) -> bool {
         self.len() == 0
     }
 
     /// The section as a slice.
+    #[inline]
     pub fn as_slice(&self) -> &[T] {
         if self.is_empty() {
             &[]
@@ -954,6 +923,7 @@ impl<T: 'static> TypedSection<T> {
     }
 
     /// The offset of the item in the section, if it is in the section.
+    #[inline]
     pub fn offset_of(&self, item: &T) -> Option<usize> {
         let ptr = item as *const T;
         if ptr < self.start_ptr() || ptr >= self.end_ptr() {
@@ -970,24 +940,13 @@ impl<T: 'static> TypedSection<T> {
     /// This cannot be safely used and is _absolutely unsound_ if any other
     /// slices are live.
     #[allow(clippy::mut_from_ref)]
+    #[inline]
     pub unsafe fn as_mut_slice(&self) -> &mut [T] {
         if self.is_empty() {
             &mut []
         } else {
             unsafe { ::core::slice::from_raw_parts_mut(self.start_ptr() as *mut T, self.len()) }
         }
-    }
-
-    /// The start address of the section.
-    #[inline(always)]
-    pub fn start_ptr_mut(&self) -> *mut T {
-        self.start_ptr() as *mut T
-    }
-
-    /// The start address of the section.
-    #[inline(always)]
-    pub fn end_ptr_mut(&self) -> *mut T {
-        self.end_ptr() as *mut T
     }
 }